feat: overall auth and date transformation

+++ client/src/app/api/subscription-plans/_service/deletePlan.service.ts
@@ -1,8 +1,20 @@
+import { CustomErrorResponse } from "@/backend/helpers/customErrorResponse";
 import prisma from "@/lib/db";
 import { NextResponse } from "next/server";
 
 export async function deletePlan(id: number) {
   console.log("THE ID IS", id);
+  const assosiatedSubscription = await prisma.subscription.findFirst({
+    where: {
+      planId: id,
+    },
+    select: {
+      id: true,
+    },
+  });
+  if (assosiatedSubscription) {
+    return CustomErrorResponse("Cannot delete plan, subscriptions exist", 400);
+  }
   const plan = await prisma.subscriptionPlan.delete({
     where: { id },
   });
